Model pitch according to twelve-tone equal temperament
Pitch is the auditory attribute of sound according to which sounds can
be ordered on a scale from low to high. In other words, it is the
property which allows one sound to be perceived as "higher" or "lower"
than another.

The new Pitch record consists of a note letter and an accidental
modifier which sharpens or flattens the note. This conforms to the
twelve-tone equal temperament used in classical music, and Western
music in general.