Allow an interval to be constructed from a semitone count
This makes assumptions about which interval to create since semitone
count alone is not enough; multiple intervals have the same count. The
most common intervals are created right now, though in future this
should be improved to consider the Number as well. For example, a count
of 6 semitones could refer to an augmented fourth or a diminished fifth.
By specifying the Number, it is possible to know which to select.