Model scale forms which adjust intervals based on a condition
A simple model is implemented in which intervals can be adjusted
always or based on the direction in which the scale is being played.
This allows for scales such as melodic minor with augmented 6th and 7th
steps.

The same mechanism can also be used to build scales such as harmonic
minor by defining an aeolian major with an augmented 7th form which
applies in either direction. This removes the need to adjust intervals
before building the scale and supports more powerful composition of
scale features.